requirements.txt
Jinja2==2.10
MarkupSafe==1.0
PyYAML==3.12
